Make two mixins.json for clarity
